+++ ecs-cli/modules/config/params.go
@@ -1,8 +1,3 @@
-// NOTE: DEPRECATED. These structs are only left here so that
-// we can read old ini based config files for customers who have
-// still been using older versions of the CLI. All new config files
-// will be written in the YAML format.
-
 // Copyright 2015-2017 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"). You may
@@ -18,21 +13,75 @@
 
 package config
 
-// oldCliConfig is the struct used to map to the ini config.
-// This is to allow us to read old ini based config files
-// CliConfig has been updated to use the yaml annotations
-type oldCliConfig struct {
-	*oldSectionKeys `ini:"ecs"`
+import (
+	"fmt"
+	"os"
+
+	"github.com/Sirupsen/logrus"
+	ecscli "github.com/aws/amazon-ecs-cli/ecs-cli/modules/commands"
+	"github.com/aws/aws-sdk-go/aws/session"
+	"github.com/urfave/cli"
+)
+
+// CliParams saves config to create an aws service clients
+type CliParams struct {
+	Cluster                  string
+	Session                  *session.Session
+	ComposeProjectNamePrefix string
+	ComposeServiceNamePrefix string
+	CFNStackNamePrefix       string
 }
 
-// SectionKeys is the struct embedded in oldCliConfig. It groups all the keys in the 'ecs' section in the ini file.
-type oldSectionKeys struct {
-	Cluster                  string `ini:"cluster"`
-	AwsProfile               string `ini:"aws_profile"`
-	Region                   string `ini:"region"`
-	AwsAccessKey             string `ini:"aws_access_key_id"`
-	AwsSecretKey             string `ini:"aws_secret_access_key"`
-	ComposeProjectNamePrefix string `ini:"compose-project-name-prefix"`
-	ComposeServiceNamePrefix string `ini:"compose-service-name-prefix"`
-	CFNStackNamePrefix       string `ini:"cfn-stack-name-prefix"`
+// GetCfnStackName <cfn_stack_name_prefix> + <cluster_name>
+func (p *CliParams) GetCfnStackName() string {
+	return fmt.Sprintf("%s%s", p.CFNStackNamePrefix, p.Cluster)
+}
+
+// NewCliParams creates a new ECSParams object from the config file.
+func NewCliParams(context *cli.Context, rdwr ReadWriter) (*CliParams, error) {
+	ecsConfig, configMap, err := rdwr.GetConfig()
+	if err != nil {
+		logrus.Error("Error loading config: ", err)
+		return nil, err
+	}
+
+	// If Prefixes not found, set to defaults.
+	if _, ok := configMap[composeProjectNamePrefixKey]; !ok {
+		ecsConfig.ComposeProjectNamePrefix = ecscli.ComposeProjectNamePrefixDefaultValue
+	}
+	if _, ok := configMap[composeServiceNamePrefixKey]; !ok {
+		ecsConfig.ComposeServiceNamePrefix = ecscli.ComposeServiceNamePrefixDefaultValue
+	}
+	if _, ok := configMap[cfnStackNamePrefixKey]; !ok {
+		ecsConfig.CFNStackNamePrefix = ecscli.CFNStackNamePrefixDefaultValue
+	}
+
+	// Order of cluster resolution
+	//  1) Inline flag
+	//  2) Environment Variable
+	//  3) ECS Config
+	if clusterFromEnv := os.Getenv(ecscli.ClusterEnvVar); clusterFromEnv != "" {
+		ecsConfig.Cluster = clusterFromEnv
+	}
+	if clusterFromFlag := context.String(ecscli.ClusterFlag); clusterFromFlag != "" {
+		ecsConfig.Cluster = clusterFromFlag
+	}
+
+	// local --region flag has the highest precedence to set ecs-cli region config.
+	if regionFromFlag := context.String(ecscli.RegionFlag); regionFromFlag != "" {
+		ecsConfig.Region = regionFromFlag
+	}
+
+	svcSession, err := ecsConfig.ToAWSSession()
+	if err != nil {
+		return nil, err
+	}
+
+	return &CliParams{
+		Cluster:                  ecsConfig.Cluster,
+		Session:                  svcSession,
+		ComposeProjectNamePrefix: ecsConfig.ComposeProjectNamePrefix,
+		ComposeServiceNamePrefix: ecsConfig.ComposeServiceNamePrefix,
+		CFNStackNamePrefix:       ecsConfig.CFNStackNamePrefix,
+	}, nil
 }
